LPS-189757 Refactoring the entire implementation and validations
