+++ modules/apps/site-initializer/site-initializer-liferay-marketplace/src/main/resources/site-initializer/commerce-catalogs/catalog.products.json
@@ -13,6 +13,7 @@
 		"externalReferenceCode": "TESTFREEPRODUCT",
 		"name": "Test Free Product",
 		"price": 0,
+		"sku": "TESTFREEPRODUCTSKU",
 		"specificationOptions": [
 		]
 	}
