+++ modules/apps/site-initializer/site-initializer-liferay-marketplace/src/main/resources/site-initializer/layouts/03_solutions_marketplace/02_purchase_product_form/page-definition.json
@@ -28,7 +28,7 @@
 							"widgetInstance": {
 								"widgetConfig": {
 								},
-								"widgetName": "com_liferay_client_extension_web_internal_portlet_ClientExtensionEntryPortlet_[$COMPANY_ID$]_LXC_liferay_marketplace_custom_element_purchased_get_app"
+								"widgetName": "com_liferay_client_extension_web_internal_portlet_ClientExtensionEntryPortlet_[$COMPANY_ID$]_LXC_liferay_marketplace_custom_element_purchased_solutions"
 							}
 						},
 						"type": "Widget"
