+++ modules/apps/site-initializer/site-initializer-liferay-marketplace/src/main/resources/site-initializer/object-definitions/userAdditionalInfo.object-actions.json
@@ -13,7 +13,7 @@
 		"objectActionExecutorKey": "update-object-entry",
 		"objectActionTriggerKey": "onAfterAdd",
 		"parameters": {
-			"objectDefinitionExternalReferenceCode": "MKT-USER-ADDITIONAL",
+			"objectDefinitionExternalReferenceCode": "USERADDITIONAL",
 			"predefinedValues": [
 				{
 					"businessType": "Text",
