+++ modules/apps/site-initializer/site-initializer-liferay-marketplace/src/main/resources/site-initializer/object-fields/company-field.json
@@ -2,7 +2,7 @@
 	"object-fields": [
 		{
 			"DBType": "String",
-			"externalReferenceCode": "MKT-COMPANY",
+			"externalReferenceCode": "COMPANY",
 			"indexed": true,
 			"indexedAsKeyword": false,
 			"label": {
