+++ modules/apps/site-initializer/site-initializer-liferay-marketplace/src/main/resources/site-initializer/object-fields/project-name-field.json
@@ -2,7 +2,7 @@
 	"object-fields": [
 		{
 			"DBType": "String",
-			"externalReferenceCode": "MKT-PROJECT-NAME",
+			"externalReferenceCode": "PROJECTNAME",
 			"indexed": true,
 			"indexedAsKeyword": false,
 			"label": {
