+++ modules/apps/site-initializer/site-initializer-liferay-marketplace/src/main/resources/site-initializer/object-fields/site-initializer-field.json
@@ -2,7 +2,7 @@
 	"object-fields": [
 		{
 			"DBType": "String",
-			"externalReferenceCode": "MKT-SITE-INITIALIZER",
+			"externalReferenceCode": "SITEINITIALIZER",
 			"indexed": true,
 			"indexedAsKeyword": false,
 			"label": {
