+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/client-extension.yaml
@@ -85,16 +85,16 @@ liferay-marketplace-custom-element-purchased-apps:
     urls:
         - "*.js"
     useESM: true
-liferay-marketplace-custom-element-purchased-get-app:
+liferay-marketplace-custom-element-purchased-solutions:
     cssURLs:
         - "**/*.css"
-    friendlyURLMapping: liferay-marketplace-custom-element-purchased-get-app
+    friendlyURLMapping: liferay-marketplace-custom-element-purchased-solutions
     htmlElementName: liferay-marketplace-custom-element
     instanceable: false
-    name: Liferay Marketplace Purchased Get app
+    name: Liferay Marketplace Purchased Solutions
     portletCategoryName: category.client-extensions
     properties:
-        route: purchased-get-app
+        route: purchased-solutions
     type: customElement
     urls:
         - "*.js"
