+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/PurchasedSolutions/CreatedProjectCard.tsx
@@ -1,5 +1,3 @@
-/* eslint-disable no-return-assign */
-/* eslint-disable @typescript-eslint/no-unused-vars */
 /**
  * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
  *
@@ -18,29 +16,21 @@ import ClayButton from '@clayui/button';
 
 import createdProjectIcon from '../../assets/images/created_project.svg';
 
-import './PurchasedGetAppPage.scss';
+import './PurchasedSolutions.scss';
 
 import ClayIcon from '@clayui/icon';
 
 import {getSiteURL} from '../../components/InviteMemberModal/services';
 import {Liferay} from '../../liferay/liferay';
 
-type Steps = {
-	page: 'accountCreation' | 'accountSelection' | 'projectCreated';
-};
-
 type CreatedProjectCardProps = {
 	product?: Product;
-	setStep: React.Dispatch<Steps>;
 };
 
-const CreatedProjectCard: React.FC<CreatedProjectCardProps> = ({
-	product,
-	setStep,
-}) => {
+const CreatedProjectCard: React.FC<CreatedProjectCardProps> = ({product}) => {
 	return (
-		<div className="align-items-center d-flex flex-column h-100 justify-content-center purchased-get-app-page-container w-100">
-			<div className="border p-8 purchased-get-app-page-body rounded">
+		<div className="align-items-center d-flex flex-column h-100 justify-content-center purchased-solutions-container w-100">
+			<div className="border p-8 purchased-solutions-body rounded">
 				<div className="align-items-center d-flex flex-column justify-content-center">
 					<div className="mb-6">
 						<img
@@ -68,12 +58,12 @@ const CreatedProjectCard: React.FC<CreatedProjectCardProps> = ({
 						</div>
 					</div>
 
-					<div className="mt-6 purchased-get-app-page-button-container">
+					<div className="mt-6 purchased-solutions-button-container">
 						<ClayButton
 							className="py-3"
-							onClick={() =>
-								(window.location.href = `${Liferay.ThemeDisplay.getPortalURL()}${getSiteURL()}/solutions-marketplace`)
-							}
+							onClick={() => {
+								window.location.href = `${Liferay.ThemeDisplay.getPortalURL()}${getSiteURL()}/solutions-marketplace`;
+							}}
 						>
 							Return to Dashboard
 							<span className="ml-3">
