+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/PurchasedSolutions/PurchasedSolutions.scss
@@ -1,5 +1,4 @@
-.purchased-get-app-page {
-	display: flex;
+.purchased-solutions {
 	margin-top: 4.5rem;
 
 	&-phone {
@@ -139,13 +138,6 @@
 		color: #54555f;
 		font-size: 16px;
 	}
-}
-
-.create-project-card {
-	align-self: stretch;
-	display: flex;
-	flex-direction: column;
-	justify-content: center;
 }
 
 .created-project-cart-title {
